requirements_higgs.txt
stream2sentence==0.3.1
phonemizer==3.3.0
huggingface-hub
loguru
dacite
tqdm
halo
scipy
aiohttp
pandas
inflect
resampy
pydantic
soundfile
safetensors
numpy==1.26.0
accelerate>=0.26.0
transformers<4.47.0
vector_quantize_pytorch
sounddevice
requests
pydub
av
asyncio
uvicorn
fastapi
pyngrok
librosa
diffusers
omegaconf
conformer
faster-whisper==1.1.1
pvporcupine==1.9.5
webrtcvad-wheels
websockets
websocket-client
openwakeword>=0.4.0
openai
supabase